LayoutTests/inspector/editor/text-editor-selection-expected.txt

This test checks the text editor selection calculation.

Chunk model
Chunk [0] 0:9 (expanded)
Chunk [1] 9:18 (collapsed)
Chunk [2] 18:27 (collapsed)
Chunk [3] 27:34 (collapsed)

Running: testSelectionAtStartOfCollapsedChunk
{"startLine":27,"startColumn":0,"endLine":27,"endColumn":0}

Running: testSelectionAtEndOfCollapsedChunk
{"startLine":33,"startColumn":2,"endLine":33,"endColumn":2}

Running: testSelectionAtMiddleOfCollapsedChunk
{"startLine":27,"startColumn":1,"endLine":30,"endColumn":2}

Running: testSelectionAtStartAndEndOfExpandedRows
{"startLine":3,"startColumn":0,"endLine":4,"endColumn":1}

Running: testSelectionAtMiddleOfExpandedRows
{"startLine":3,"startColumn":1,"endLine":4,"endColumn":1}
